Fix severe bug
Fixed a severe bug that would cause the program to crash when writing the perfect amount of data to a slice. Ex: I want to write an int32 to the last 4 bytes of a slice, if I performed the WriteInt32 or and int32 write operation the program would crash with an index out of bounds error. I do not know if there will be an overall performance hit or not but I'm sure it will be nearly negligible.